+++ tests-ui/tests/litegraph/core/LGraphGroup.test.ts
@@ -2,7 +2,7 @@ import { describe, expect } from 'vitest'
 
 import { LGraphGroup } from '@/lib/litegraph/src/litegraph'
 
-import { test } from '../fixtures/testExtensions'
+import { test } from './fixtures/testExtensions'
 
 describe('LGraphGroup', () => {
   test('serializes to the existing format', () => {
