+++ tests-ui/tests/litegraph/core/fixtures/assets/floatingLink.json
