--- tests-ui/tests/litegraph/fixtures/assets/linkedNodes.json
@@ -1,96 +1,0 @@
-{
-  "id": "26a34f13-1767-4847-b25f-a21dedf6840d",
-  "revision": 0,
-  "last_node_id": 3,
-  "last_link_id": 2,
-  "nodes": [
-    {
-      "id": 2,
-      "type": "VAEDecode",
-      "pos": [
-        63.44815444946289,
-        178.71633911132812
-      ],
-      "size": [
-        210,
-        46
-      ],
-      "flags": {},
-      "order": 0,
-      "mode": 0,
-      "inputs": [
-        {
-          "name": "samples",
-          "type": "LATENT",
-          "link": null
-        },
-        {
-          "name": "vae",
-          "type": "VAE",
-          "link": null
-        }
-      ],
-      "outputs": [
-        {
-          "name": "IMAGE",
-          "type": "IMAGE",
-          "links": [
-            2
-          ]
-        }
-      ],
-      "properties": {
-        "Node name for S&R": "VAEDecode"
-      },
-      "widgets_values": []
-    },
-    {
-      "id": 3,
-      "type": "SaveImage",
-      "pos": [
-        419.36920166015625,
-        179.71388244628906
-      ],
-      "size": [
-        226.3714141845703,
-        58
-      ],
-      "flags": {},
-      "order": 1,
-      "mode": 0,
-      "inputs": [
-        {
-          "name": "images",
-          "type": "IMAGE",
-          "link": 2
-        }
-      ],
-      "outputs": [],
-      "properties": {},
-      "widgets_values": [
-        "ComfyUI"
-      ]
-    }
-  ],
-  "links": [
-    [
-      2,
-      2,
-      0,
-      3,
-      0,
-      "IMAGE"
-    ]
-  ],
-  "groups": [],
-  "config": {},
-  "extra": {
-    "linkExtensions": [
-      {
-        "id": 2,
-        "parentId": 1
-      }
-    ]
-  },
-  "version": 0.4
-}
