+++ tests-ui/tests/litegraph/subgraph/SubgraphNode.test.ts
@@ -9,11 +9,11 @@ import { describe, expect, it, vi } from 'vitest'
 
 import { LGraph, Subgraph } from '@/lib/litegraph/src/litegraph'
 
-import { subgraphTest } from '../../fixtures/subgraphFixtures'
+import { subgraphTest } from './fixtures/subgraphFixtures'
 import {
   createTestSubgraph,
   createTestSubgraphNode
-} from '../../fixtures/subgraphHelpers'
+} from './fixtures/subgraphHelpers'
 
 describe.skip('SubgraphNode Construction', () => {
   it('should create a SubgraphNode from a subgraph definition', () => {
